Updated the contents with a few changes. Please see the description
I have updated some of the page contents. Also I created buttons to direct users to the product doc page(s). Its just a guide as to what is expected. Please go ahead with styling the buttons.

We need the team section not to be in slider. We need all the images in a single container and each image in a rounded form. If you need more direction, please ask on the forum.